秒杀商品上架

+++ gulimall-seckill/src/main/java/top/ctong/gulimall/seckill/to/SeckillSkuRelationTo.java
@@ -47,11 +47,11 @@ public class SeckillSkuRelationTo {
     /**
      * 秒杀总量
      */
-    private BigDecimal seckillCount;
+    private Integer seckillCount;
     /**
      * 每人限购数量
      */
-    private BigDecimal seckillLimit;
+    private Integer seckillLimit;
     /**
      * 排序
      */
